+++ packages/flat-server-api/src/room.ts
@@ -1,5 +1,5 @@
 import { RoomStatus, RoomType, Week, Region } from "./constants";
-import { post, post2, postNotAuth, PROCESSING } from "./utils";
+import { post } from "./utils";
 
 export interface CreateOrdinaryRoomPayload {
     title: string;
@@ -388,108 +388,6 @@ export function usersInfo(payload: UsersInfoPayload): Promise<UsersInfoResult> {
     return post<UsersInfoPayload, UsersInfoResult>("room/info/users", payload);
 }
 
-export interface LoginCheckPayload {}
-
-export interface LoginCheckResult {
-    name: string;
-    avatar: string;
-    token: string;
-    userUUID: string;
-    hasPhone: boolean;
-}
-
-export async function loginCheck(token?: string): Promise<LoginCheckResult> {
-    return await post<LoginCheckPayload, LoginCheckResult>("login", {}, {}, token);
-}
-
-export interface SetAuthUUIDPayload {
-    authUUID: string;
-}
-
-export interface SetAuthUUIDResult {
-    authUUID: string;
-}
-
-export async function setAuthUUID(authUUID: string): Promise<SetAuthUUIDResult> {
-    return await postNotAuth<SetAuthUUIDPayload, SetAuthUUIDResult>("login/set-auth-uuid", {
-        authUUID,
-    });
-}
-
-export interface LoginProcessPayload {
-    authUUID: string;
-}
-
-export interface LoginProcessResult {
-    name: string;
-    avatar: string;
-    userUUID: string;
-    token: string;
-    hasPhone: boolean;
-}
-
-export async function loginProcess(authUUID: string): Promise<LoginProcessResult> {
-    return await postNotAuth<LoginProcessPayload, LoginProcessResult>("login/process", {
-        authUUID,
-    });
-}
-
-export interface LoginPhoneSendCodePayload {
-    phone: string; // +8612345678901
-}
-
-export type LoginPhoneSendCodeResult = {};
-
-export async function loginPhoneSendCode(phone: string): Promise<LoginPhoneSendCodeResult> {
-    return await postNotAuth<LoginPhoneSendCodePayload, LoginPhoneSendCodeResult>(
-        "login/phone/sendMessage",
-        {
-            phone,
-        },
-    );
-}
-
-export interface LoginPhonePayload {
-    phone: string; // +8612345678901
-    code: number; // 123456
-}
-
-export async function loginPhone(phone: string, code: number): Promise<LoginProcessResult> {
-    return await postNotAuth<LoginPhonePayload, LoginProcessResult>("login/phone", {
-        phone,
-        code,
-    });
-}
-
-export interface BindingPhoneSendCodePayload {
-    phone: string; // +8612345678901
-}
-
-export type BindingPhoneSendCodeResult = {};
-
-export async function bindingPhoneSendCode(phone: string): Promise<BindingPhoneSendCodeResult> {
-    return await post<BindingPhoneSendCodePayload, BindingPhoneSendCodeResult>(
-        "user/binding/platform/phone/sendMessage ",
-        {
-            phone,
-        },
-    );
-}
-
-export interface BindingPhonePayload {
-    phone: string; // +8612345678901
-    code: number; // 123456
-}
-
-export type BindingPhoneResult = {};
-
-export async function bindingPhone(phone: string, code: number): Promise<BindingPhoneResult> {
-    return await post<BindingPhonePayload, BindingPhoneResult>("user/binding/platform/phone", {
-        phone,
-        code,
-    });
-}
-
 export interface RenamePayload {
     name: string;
 }
@@ -543,90 +441,4 @@ export async function uploadAvatarFinish(fileUUID: string): Promise<UploadAvatar
             fileUUID,
         },
     );
-}
-
-export interface ListBindingsPayload {}
-
-export interface ListBindingsResult {
-    wechat: boolean;
-    phone: boolean;
-    agora: boolean;
-    apple: boolean;
-    github: boolean;
-    google: boolean;
-}
-
-export async function listBindings(): Promise<ListBindingsResult> {
-    return await post<ListBindingsPayload, ListBindingsResult>("user/binding/list", {});
-}
-
-export interface SetBindingAuthUUIDResult {}
-
-export async function setBindingAuthUUID(authUUID: string): Promise<void> {
-    await post<SetAuthUUIDPayload, SetBindingAuthUUIDResult>("user/binding/set-auth-uuid", {
-        authUUID,
-    });
-}
-
-export interface BindingProcessResult {
-    processing: boolean;
-    status: boolean;
-}
-
-export async function bindingProcess(authUUID: string): Promise<BindingProcessResult> {
-    try {
-        const ret = await post2<SetAuthUUIDPayload, {}>("user/binding/process", {
-            authUUID,
-        });
-        if (ret === PROCESSING) {
-            return {
-                processing: true,
-                status: false,
-            };
-        }
-        return {
-            processing: false,
-            status: true,
-        };
-    } catch {
-        return {
-            processing: false,
-            status: false,
-        };
-    }
-}
-
-export enum LoginPlatform {
-    WeChat = "WeChat",
-    Github = "Github",
-    Apple = "Apple",
-    Agora = "Agora",
-    Google = "Google",
-    Phone = "Phone",
-}
-
-export interface RemoveBindingPayload {
-    target: LoginPlatform;
-}
-
-export interface RemoveBindingResult {
-    token: string;
-}
-
-export async function removeBinding(target: LoginPlatform): Promise<RemoveBindingResult> {
-    return await post<RemoveBindingPayload, RemoveBindingResult>("user/binding/remove", {
-        target,
-    });
-}
-
-export interface DeleteAccountValidateResult {
-    alreadyJoinedRoomCount: number;
-}
-
-export async function deleteAccountValidate(): Promise<DeleteAccountValidateResult> {
-    return await post<{}, DeleteAccountValidateResult>("user/deleteAccount/validate", {});
-}
-
-export async function deleteAccount(): Promise<void> {
-    await post<{}, {}>("user/deleteAccount", {});
 }
