+++ packages/flat-stores/package.json
@@ -12,20 +12,20 @@
     "@netless/flat-rtc": "workspace:*",
     "@netless/flat-rtm": "workspace:*",
     "@netless/flat-server-api": "workspace:*",
-    "@netless/synced-store": "1.x",
     "emittery": "0.x",
     "flat-components": "workspace:*",
     "flat-i18n": "workspace:*",
     "mobx": "6.x",
     "side-effect-manager": "^0.1.5",
     "uuid": "8.x",
-    "white-web-sdk": ">= 2.16"
+    "white-web-sdk": ">=2.16"
   },
   "devDependencies": {
+    "@netless/fastboard-core": "^0.3.5-canary.2",
     "@netless/flat-rtc": "workspace:*",
     "@netless/flat-rtm": "workspace:*",
     "@netless/flat-server-api": "workspace:*",
-    "@netless/synced-store": "^1.1.1",
+    "@netless/window-manager": "^0.4.31",
     "emittery": "^0.10.2",
     "flat-components": "workspace:*",
     "flat-i18n": "workspace:*",
@@ -34,6 +34,6 @@
     "side-effect-manager": "^0.1.5",
     "typescript": "^4.6.2",
     "uuid": "^8.3.2",
-    "white-web-sdk": "2.16.20"
+    "white-web-sdk": "npm:white-web-sdk-esm@2.16.24"
   }
 }
