+++ web/flat-web/package.json
@@ -21,7 +21,7 @@
     "@netless/app-plyr": "^0.1.3",
     "@netless/app-slide": "^0.2.2",
     "@netless/combine-player": "^1.1.6",
-    "@netless/fastboard-react": "^0.3.5-canary.1",
+    "@netless/fastboard-react": "^0.3.5-canary.2",
     "@netless/flat-rtc": "workspace:*",
     "@netless/flat-rtc-agora-web": "workspace:*",
     "@netless/flat-rtm": "workspace:*",
