+++ web/flat-web/src/pages/HomePage/MainRoomListPanel/MainRoomList.tsx
@@ -13,13 +13,17 @@ import {
     RoomStatusType,
     errorTips,
 } from "flat-components";
-import { ListRoomsType, RoomStatus, RoomType } from "@netless/flat-server-api";
+import {
+    FLAT_SERVER_BASE_URL,
+    ListRoomsType,
+    RoomStatus,
+    RoomType,
+} from "@netless/flat-server-api";
 import { GlobalStoreContext, RoomStoreContext } from "../../../components/StoreProvider";
 import { RoomItem } from "@netless/flat-stores";
 import { useSafePromise } from "../../../utils/hooks/lifecycle";
 import { RouteNameType, usePushHistory } from "../../../utils/routes";
 import { joinRoomHandler } from "../../utils/join-room-handler";
-import { FLAT_WEB_BASE_URL } from "../../../constants/process";
 import { useTranslation } from "react-i18next";
 
 export interface MainRoomListProps {
@@ -229,7 +233,7 @@ export const MainRoomList = observer<MainRoomListProps>(function MainRoomList({
             )}
             {currentRoom && (
                 <InviteModal
-                    baseUrl={FLAT_WEB_BASE_URL}
+                    baseUrl={FLAT_SERVER_BASE_URL}
                     periodicWeeks={periodicInfo?.periodic.weeks}
                     room={currentRoom}
                     userName={globalStore.userName ?? ""}
@@ -259,7 +263,10 @@ export const MainRoomList = observer<MainRoomListProps>(function MainRoomList({
 
     function replayRoom(config: { roomUUID: string; ownerUUID: string; roomType: RoomType }): void {
         const { roomUUID, ownerUUID, roomType } = config;
-        window.open(`${FLAT_WEB_BASE_URL}/replay/${roomType}/${roomUUID}/${ownerUUID}/`, "_blank");
+        window.open(
+            `${FLAT_SERVER_BASE_URL}/replay/${roomType}/${roomUUID}/${ownerUUID}/`,
+            "_blank",
+        );
     }
 
     function hideCancelModal(): void {
